+++ wiswa/tool/__init__.py
@@ -1,4 +1,4 @@
 """wiswa module."""
 from __future__ import annotations
 
-__version__ = '0.6.1'
+__version__ = '0.6.2'